revert folder restructuring
