+++ src/components/Loader/style.module.css
