+++ src/components/index.ts
@@ -1,13 +1,13 @@
-export * from './global/Button';
-export * from './global/Input';
-export * from './global/Loader';
-export * from './global/Modal';
-export * from './global/PasswordInput';
-export * from './layout/Header';
-export * from './layout/Footer';
-export * from './layout/Nav';
-export * from './layout/NavItem';
-export * from './layout/UserList';
-export * from './layout/DashboardAdmin';
-export * from './layout/DashboardUser';
-export * from './layout/UserCard';
+export * from './Button';
+export * from './Input';
+export * from './Loader';
+export * from './Modal';
+export * from './PasswordInput';
+export * from './Header';
+export * from './Footer';
+export * from './Nav';
+export * from './NavItem';
+export * from './UserList';
+export * from './DashboardAdmin';
+export * from './DashboardUser';
+export * from './UserCard';
